Add dotenv configuration split the code in different files
