+++ .env
@@ -1,0 +1,2 @@
+MAIN_DIR=/home/alpha/Desktop/eos
+PORT=3001
