+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.5.0",
     "express": "^5.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
